Move truncateClazzes to dbutil
