+++ package.json
@@ -18,6 +18,7 @@
     "fast-csv": "^4.3.6",
     "inflection": "^1.12.0",
     "jsonwebtoken": "^8.5.1",
+    "lodash.clonedeep": "^4.5.0",
     "mime-types": "^2.1.27",
     "multer": "^1.4.2",
     "pg": "^8.4.1",
@@ -34,6 +35,7 @@
     "@types/inflection": "^1.5.28",
     "@types/jest": "^26.0.15",
     "@types/jsonwebtoken": "^8.5.0",
+    "@types/lodash.clonedeep": "^4.5.6",
     "@types/mime-types": "^2.1.0",
     "@types/multer": "^1.4.5",
     "@types/node": "^14.11.2",
