+++ src/server.ts
@@ -3,7 +3,7 @@ import { Request, Response, NextFunction } from 'express';
 import cors from 'cors';
 
 import { PORT } from './consts';
-// import {testAuthenticate} from './db/dbUtil';
+// import { testAuthenticate } from './utils/dbUtil';
 import './models'; // import for side effects
 
 import { HTTPError, NotFoundError } from './errors/httpErrors';
